Improve: GoogleCalendarApiコントローラの処理の一部をServicesクラスに切り分け
