--- app/services/google_calendar/service.rb
@@ -1,2 +1,0 @@
-class GoogleCalendar::Service
-end
